+++ apps/main-app/ts/features/lollipop/utils/__tests__/retryAndVerifyMissingSamlRequest.test.ts
@@ -1,4 +1,3 @@
-import * as E from "fp-ts/lib/Either";
 import { PublicKey } from "@pagopa/io-react-native-crypto";
 import { regenerateKeyGetRedirectsAndVerifySaml } from "../login";
 import { AppDispatch } from "../../../../App";
@@ -25,17 +24,15 @@ jest.mock("@pagopa/io-react-native-login-utils", () => ({
 }));
 
 describe("Lollipop regenerate key, get redirects and verification", () => {
-  it("should be succeded", async () => {
-    const result = await regenerateKeyGetRedirectsAndVerifySaml(
-      "loginUri",
-      "keyTag",
-      false,
-      false,
-      dispatch
-    );
-    expect(E.isLeft(result)).toBeTruthy();
-    if (E.isLeft(result)) {
-      expect(result.left).toEqual(new Error("Missing SAMLRequest"));
-    }
+  it("should reject because SAMLRequest is missing", async () => {
+    await expect(
+      regenerateKeyGetRedirectsAndVerifySaml(
+        "loginUri",
+        "keyTag",
+        false,
+        false,
+        dispatch
+      )
+    ).rejects.toEqual(new Error("Missing SAMLRequest"));
   });
 });