+++ apps/main-app/ts/features/pn/aar/store/selectors/__tests__/index.test.ts
@@ -84,8 +84,8 @@ describe("thirdPartySenderDenominationSelector", () => {
     expect(result).toBeUndefined();
   });
 });
-describe(" currentAarFlowData and currentAarFlowStateType", () => {
-  it(" should return the correct AAR flow state from the global state", () => {
+describe("currentAarFlowData and currentAarFlowStateType", () => {
+  it("should return the correct AAR flow state from the global state", () => {
     const mockFetchingState =
       sendAarMockStateFactory.fetchingNotificationData();